requirements.txt
Django~=4.2
python-dotenv~=1.0.0
environs==9.5.0
aiogram~=2.25.1
asgiref~=3.6.0
requests~=2.28.2
phonenumberslite==8.13.7
django-phonenumber-field==7.1.0
django-admin-rangefilter==0.10.0